+++ lib/config/all.dart
@@ -1,3 +1,4 @@
 export 'configs.dart';
+export 'custom_selection_config.dart';
 export 'markdown_generator.dart';
 export 'toc.dart';